Project Restructuring pt.3 [Beta 0.15] -Moved Most Function code to `common` -Generated and data-driven objects moved to `datagen`
