+++ src/main/java/me/shadowalzazel/mcodyssey/common/listeners/SoulBraisingListener.kt
@@ -1,7 +1,7 @@
 package me.shadowalzazel.mcodyssey.common.listeners
 
 import me.shadowalzazel.mcodyssey.Odyssey
-import me.shadowalzazel.mcodyssey.alchemy.SoulBraiseRecipes
+import me.shadowalzazel.mcodyssey.common.alchemy.SoulBraiseRecipes
 import me.shadowalzazel.mcodyssey.util.constants.ItemModels
 import me.shadowalzazel.mcodyssey.util.constants.ItemDataTags
 import me.shadowalzazel.mcodyssey.common.items.custom.Ingredients
