+++ src/main/java/me/shadowalzazel/mcodyssey/util/DataKeys.kt
@@ -1,4 +1,4 @@
-package me.shadowalzazel.mcodyssey.util.constants
+package me.shadowalzazel.mcodyssey.util
 
 import me.shadowalzazel.mcodyssey.Odyssey
 import org.bukkit.NamespacedKey
@@ -7,5 +7,7 @@ object DataKeys {
 
     val ITEM_KEY = NamespacedKey(Odyssey.instance, "item")
     val UUID_KEY = NamespacedKey(Odyssey.instance, "UUID")
+
+    val newKey: (String) -> NamespacedKey = { key: String -> NamespacedKey(Odyssey.instance, key) }
 
 }
