+++ src/main/java/me/shadowalzazel/mcodyssey/util/DataTagManager.kt
@@ -1,7 +1,6 @@
 package me.shadowalzazel.mcodyssey.util
 
 import me.shadowalzazel.mcodyssey.Odyssey
-import me.shadowalzazel.mcodyssey.util.constants.DataKeys
 import org.bukkit.NamespacedKey
 import org.bukkit.inventory.ItemStack
 import org.bukkit.persistence.PersistentDataContainer
